requirements.txt
discord.py
urbandictionary
pyfiglet
bs4
lxml
youtube_dl